+++ build.sbt
@@ -116,8 +116,8 @@ lazy val core = scalajsProject("core")
     libraryDependencies ++= Seq(
       "com.softwaremill.magnolia1_3" %%% "magnolia" % "1.3.3",
       "com.raquo" %%% "laminar" % "16.0.0",
-      "io.laminext" %%% "websocket" % "0.16.2",
       "be.doeraene" %%% "web-components-ui5" % "1.17.0",
+      "io.laminext" %%% "websocket" % "0.16.2",
       "io.github.iltotore" %%% "iron" % "2.3.0"
     )
   )
@@ -133,13 +133,10 @@ lazy val ui5 = scalajsProject("ui5")
     }
   )
   .settings(scalacOptions ++= usedScalacOptions)
+  .dependsOn(core)
   .settings(
     libraryDependencies ++= Seq(
-      "com.softwaremill.magnolia1_3" %%% "magnolia" % "1.3.3",
-      "com.raquo" %%% "laminar" % "16.0.0",
-      "io.laminext" %%% "websocket" % "0.16.2",
-      "be.doeraene" %%% "web-components-ui5" % "1.17.0",
-      "io.github.iltotore" %%% "iron" % "2.3.0"
+      "be.doeraene" %%% "web-components-ui5" % "1.17.0"
     )
   )
 
@@ -153,7 +150,7 @@ lazy val example = scalajsProject("example-client", Some("example/client"))
     }
   )
   .settings(scalacOptions ++= usedScalacOptions)
-  .dependsOn(core, sharedJs)
+  .dependsOn(ui5, sharedJs)
 
 lazy val shared = crossProject(JSPlatform, JVMPlatform)
   .crossType(CrossType.Pure)
